refact: code refactor
#82
