+++ inference.py
@@ -9,49 +9,8 @@ from torch.utils.data import DataLoader
 import numpy as np
 import torch.nn as nn
 
-from dataset import CustomDataset
-import json
-
-def load_model(saved_model, num_classes, device,json_data):
-    """_summary_
-
-    Args:
-        saved_model (str): saved_model path
-        num_classes (int): 모델 구조를 동일하게 맞추기 위한 num_classes
-        device (str): device 
-
-    Returns:
-        model : weight를 불러온 모델 
-    """
-
-    
-    model_module = getattr(import_module("model"), json_data["model"])  # config.json 에 있는 파일
-    model = model_module(
-        num_classes=num_classes,
-    )
-    if json_data['canny']:
-        backbone = model
-        model_module = getattr(import_module("model"), "Canny")
-        model = model_module(
-            backbone = backbone,
-        )
-    elif json_data['arcface']:
-        backbone = model_module(num_classes=1000)
-        model_module = getattr(import_module("model"), "ArcfaceModelInfer")
-        model = model_module(
-            backbone = backbone,
-            num_features = 1000,
-            num_classes = num_classes
-        )    
-
-    # tarpath = os.path.join(saved_model, 'best.tar.gz')
-    # tar = tarfile.open(tarpath, 'r:gz')
-    # tar.extractall(path=saved_model)
-    model_path = os.path.join(saved_model, 'best.pth')
-    model.load_state_dict(torch.load(model_path, map_location=device))
-
-    return model
-
+from .datasets.my_dataset import CustomDataset
+from .utils.util import read_json, load_model, get_numclass
 
 @torch.no_grad()
 def inference(data_dir, model_dir, output_dir, args):
@@ -65,7 +24,7 @@ def inference(data_dir, model_dir, output_dir, args):
     info = pd.read_csv(info_path)
 
     img_paths = [os.path.join(img_root, img_id) for img_id in info.ImageID]
-    transform_cls = getattr(import_module("dataset"), args.augmentation)
+    transform_cls = getattr(import_module("datasets/augmentation"), args.augmentation)
     transform = transform_cls(
         resize = args.resize,
     )
@@ -101,17 +60,11 @@ def inference(data_dir, model_dir, output_dir, args):
                 }
 
             for model_path in dirlist:
-                config_path = os.path.join(model_path, 'config.json')
-                with open(config_path, 'r') as f:
-                    json_data = json.load(f)
+                json_data = read_json(os.path.join(model_path, 'config.json'))
+                
                 category = json_data['category']
 
-                if "gender" in category:
-                    num_classes = 2
-                elif "multi" in category:
-                    num_classes = 18
-                else:
-                    num_classes = 3
+                num_classes = get_numclass(category)
 
                 model = load_model(model_path, num_classes, device,json_data).to(device) 
                 model.eval()
@@ -129,6 +82,7 @@ def inference(data_dir, model_dir, output_dir, args):
                 gender = np.argmax(vote['gender'],axis=-1)
 
                 pred = mask * 6 + gender * 3 + age
+                
             preds.extend(list(pred))
 
     info['ans'] = preds
