+++ optimizers/optims.py
@@ -1,0 +1,35 @@
+from ._util import optimizer_entrypoint, is_optimizer
+
+
+def create_optimizer(optimizer_name, parameter, args):
+    """_summary_
+
+    Args:
+        optimizer_name (str): [sgd, adagrad, adam] 사용가능
+
+    Raises:
+        RuntimeError: 해당 하는 optimizer가 없다면 raise error
+
+    Returns:
+        optimizer (Module): 해당 하는 optimizer return
+    """
+    if is_optimizer(optimizer_name):
+        create_fn = optimizer_entrypoint(optimizer_name)
+        if optimizer_name in 'sgd':
+            optimizer = create_fn(
+                parameter,
+                lr= args.lr,
+                momentum = args.momentum,
+                weight_decay = args.weight_decay
+                )
+
+        else:
+            optimizer = create_fn(
+                parameter,
+                lr= args.lr,
+                weight_decay = args.weight_decay
+                )
+    else:
+        raise RuntimeError('Unknown optimizer (%s)' % optimizer_name)
+    return optimizer
+
