+++ utils/util.py
@@ -1,0 +1,169 @@
+import random
+import re
+import glob
+from pathlib import Path
+import matplotlib.pyplot as plt
+import numpy as np
+import torch
+import json
+import os
+
+from ..datasets.base_dataset import MaskBaseDataset
+from importlib import import_module
+
+
+IMG_EXTENSIONS = [
+    ".jpg", ".JPG", ".jpeg", ".JPEG", ".png",
+    ".PNG", ".ppm", ".PPM", ".bmp", ".BMP",
+]
+
+
+def is_image_file(filename):
+    return any(filename.endswith(extension) for extension in IMG_EXTENSIONS)
+
+
+def read_json(path):
+    with open(path, 'r') as f:
+        json_data = json.load(f)
+    return json_data
+
+
+def get_numclass(category):
+    if category == "multi":
+        num_classes = 18 # 18
+    elif category == "gender":
+        num_classes = 2
+    else:
+        num_classes = 3
+    return num_classes
+
+
+def get_lr(optimizer):
+    for param_group in optimizer.param_groups:
+        return param_group['lr']
+
+
+def rand_bbox(size, lam):
+    W = size[2]
+    H = size[3]
+    cut_rat = np.sqrt(1. - lam)
+    cut_w = np.int(W * cut_rat)
+    cut_h = np.int(H * cut_rat)
+
+    # uniform
+    cx = np.random.randint(W)
+    cy = np.random.randint(H)
+
+    bbx1 = np.clip(cx - cut_w // 2, 0, W)
+    bby1 = np.clip(cy - cut_h // 2, 0, H)
+    bbx2 = np.clip(cx + cut_w // 2, 0, W)
+    bby2 = np.clip(cy + cut_h // 2, 0, H)
+
+    return bbx1, bby1, bbx2, bby2
+
+
+def make_cutmix_input(inputs,labels):
+    lam = np.random.beta(1.0, 1.0)
+    rand_index = torch.randperm(inputs.size()[0]).to(inputs.device)
+    labels_a = labels
+    labels_b = labels[rand_index]
+    bbx1, bby1, bbx2, bby2 = rand_bbox(inputs.size(), lam)
+    inputs[:, :, bbx1:bbx2, bby1:bby2] = inputs[rand_index, :, bbx1:bbx2, bby1:bby2]
+    # adjust lambda to exactly match pixel ratio
+    lam = 1 - ((bbx2 - bbx1) * (bby2 - bby1) / (inputs.size()[-1] * inputs.size()[-2]))
+    return inputs, labels_a, labels_b, lam
+
+
+def seed_everything(seed):
+    torch.manual_seed(seed)
+    torch.cuda.manual_seed(seed)
+    torch.cuda.manual_seed_all(seed)  # if use multi-GPU
+    torch.backends.cudnn.deterministic = True
+    torch.backends.cudnn.benchmark = False
+    np.random.seed(seed)
+    random.seed(seed)
+
+
+def grid_image(np_images, gts, preds, n=16, shuffle=False):
+    batch_size = np_images.shape[0]
+    assert n <= batch_size
+
+    choices = random.choices(range(batch_size), k=n) if shuffle else list(range(n))
+    figure = plt.figure(figsize=(12, 18 + 2))  # cautions: hardcoded, 이미지 크기에 따라 figsize 를 조정해야 할 수 있습니다. T.T
+    plt.subplots_adjust(top=0.8)  # cautions: hardcoded, 이미지 크기에 따라 top 를 조정해야 할 수 있습니다. T.T
+    n_grid = int(np.ceil(n ** 0.5))
+    tasks = ["mask", "gender", "age"]
+    for idx, choice in enumerate(choices):
+        gt = gts[choice].item()
+        pred = preds[choice].item()
+        image = np_images[choice]
+        gt_decoded_labels = MaskBaseDataset.decode_multi_class(gt)
+        pred_decoded_labels = MaskBaseDataset.decode_multi_class(pred)
+        title = "\n".join([
+            f"{task} - gt: {gt_label}, pred: {pred_label}"
+            for gt_label, pred_label, task
+            in zip(gt_decoded_labels, pred_decoded_labels, tasks)
+        ])
+
+        plt.subplot(n_grid, n_grid, idx + 1, title=title)
+        plt.xticks([])
+        plt.yticks([])
+        plt.grid(False)
+        plt.imshow(image, cmap=plt.cm.binary)
+
+    return figure
+
+
+def increment_path(path, exist_ok=False):
+    """ Automatically increment path, i.e. runs/exp --> runs/exp0, runs/exp1 etc.
+
+    Args:
+        path (str or pathlib.Path): f"{model_dir}/{args.name}".
+        exist_ok (bool): whether increment path (increment if False).
+    """
+    path = Path(path)
+    if (path.exists() and exist_ok) or (not path.exists()):
+        return str(path)
+    else:
+        dirs = glob.glob(f"{path}*")
+        matches = [re.search(rf"%s(\d+)" % path.stem, d) for d in dirs]
+        i = [int(m.groups()[0]) for m in matches if m]
+        n = max(i) + 1 if i else 2
+        return f"{path}{n}"
+    
+
+def load_model(saved_model, num_classes, device,json_data):
+    """_summary_
+
+    Args:
+        saved_model (str): saved_model path
+        num_classes (int): 모델 구조를 동일하게 맞추기 위한 num_classes
+        device (str): device 
+
+    Returns:
+        model : weight를 불러온 모델 
+    """
+    model_module = getattr(import_module("../models/my_model"), json_data["model"])  # config.json 에 있는 파일
+    model = model_module(
+        num_classes=num_classes,
+    )
+    if json_data['canny']:
+        backbone = model
+        model_module = getattr(import_module("../models/my_model"), "Canny")
+        model = model_module(
+            backbone = backbone,
+        )
+    elif json_data['arcface']:
+        backbone = model_module(num_classes=1000)
+        model_module = getattr(import_module("../models/my_model"), "ArcfaceModelInfer")
+        model = model_module(
+            backbone = backbone,
+            num_features = 1000,
+            num_classes = num_classes
+        )    
+
+    model_path = os.path.join(saved_model, 'best.pth')
+    model.load_state_dict(torch.load(model_path, map_location=device))
+
+    return model
+
